pyproject.toml
[tool.poetry]
name = "usbserial"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyserial = "^3.5"
pyserial-asyncio = "^0.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-watch = "^4.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
